Update dashboard.py
Preprocessing of product id, sku id, buyercountry/countryofbuyer, buyerpostcode/postalcodeofbuyer. The product id and sku id are filtered for teh ddfive app. Country and Postcode had different names in a couple files so to avoid error we first check which name the current filepath(csvfile) had and stored it in a variable and then called the variable in the df function.